Login Function
Please just check had to add correct string because if I use return true or false in the for loop the method doesnt catch it as boolean
